Fix a casting bugs when it tries to laod more than 4G sized index file.

+++ src/main/java/org/opensearch/knn/index/store/IndexInputWithBuffer.java
@@ -34,7 +34,7 @@ public class IndexInputWithBuffer {
      * @throws IOException
      */
     private int copyBytes(long nbytes) throws IOException {
-        final int readBytes = Math.min(Math.toIntExact(nbytes), buffer.length);
+        final int readBytes = (int) Math.min(nbytes, buffer.length);
         indexInput.readBytes(buffer, 0, readBytes);
         return readBytes;
     }
